updated application to use pinia store instead of vuex store
