--- src/store/main.js
@@ -1,8 +1,0 @@
-import { defineStore } from "pinia";
-import state from "./state";
-import actions from "./mainactions";
-
-export const useStore = defineStore("main", {
-  state: () => state,
-  actions: actions,
-});
