+++ test/jest/__tests__/App.spec.js
@@ -2,8 +2,8 @@
 /**
  * @jest-environment jsdom
  */
-import mutations from "../../../src/store/mutations";
-import actions from "../../../src/store/actions";
+import mutations from "../../../src/store/options/mutations";
+import actions from "../../../src/store/options/actions";
 import { mount, createLocalVue, shallowMount } from "@vue/test-utils";
 import * as All from "quasar";
 const { Quasar, date } = All;
@@ -27,22 +27,22 @@ describe("Delete state/actions in userStore/componentMap", () => {
           htmlList: [],
           componentActions: [],
           state: ["state1", "state2"],
-          actions: ["action1", "action2"]
-        }
+          actions: ["action1", "action2"],
+        },
       },
       activeComponent: "test",
-      userState: ['state1', 'state2'],
-      userActions: ['action1', 'action2'],
-      userProps: ['prop1', 'prop2']
+      userState: ["state1", "state2"],
+      userActions: ["action1", "action2"],
+      userProps: ["prop1", "prop2"],
     };
   });
   test("'[types.DELETE_USER_STATE]' should delete a single state property", () => {
     mutations.DELETE_USER_STATE(state, "state1");
-    expect(state.userState).toEqual(['state2']);
+    expect(state.userState).toEqual(["state2"]);
   });
   test("'[types.DELETE_USER_ACTIONS]' should delete a single action property", () => {
     mutations.DELETE_USER_ACTIONS(state, "action1");
-    expect(state.userActions).toEqual(['action2']);
+    expect(state.userActions).toEqual(["action2"]);
   });
   test("'[types.REMOVE_ALL_STATE_PROPS_ACTIONS]' should delete all props, state and action", () => {
     mutations.REMOVE_ALL_STATE_PROPS_ACTIONS(state);
@@ -50,6 +50,4 @@ describe("Delete state/actions in userStore/componentMap", () => {
     expect(state.userActions.length).toEqual(0);
     expect(state.userProps.length).toEqual(0);
   });
-
 });
-
